Test addition of an article into Pocket
Creates a crude test and a command to execute all tests available in the
folder to mimic PHPUnit. Scopes of tests are mixed and global, saved by
their execution inside a function.

A good next step is to refactor the code to move commands into
functions, enabling them to be reused in another tests so we can create
a test that saves an article, filters and tags it.